--- genestorian_data_refinement/__init__.py
@@ -1,4 +1,0 @@
-from genestorian_data_refinement.pattern_identify import (Analyse, count_most_common_consecutive_N_characters, find)
-from genestorian_data_refinement.converge import converge
-from genestorian_data_refinement.matched_pattern import Matched_Patterns
-__all__ = ['Analyse', 'count_most_common_consecutive_N_characters', 'find']
