+++ genestorian_module/genestorian_module/__init__.py
@@ -1,0 +1,7 @@
+from genestorian_data_refinement.matched_pattern import Matched_Patterns
+from genestorian_data_refinement.converge import converge
+from genestorian_data_refinement.pattern_identify import (
+    Analyse, count_most_common_consecutive_N_characters, find)
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
