+++ genestorian_module/setup.py
@@ -1,0 +1,2 @@
+from setuptools import setup, find_packages
+setup(name='genestorian_data_refinement', packages=find_packages())
